\name{npROCRegression-package}
\alias{npROCRegression-package}
\alias{npROCRegression}
\docType{package}
\title{
Kernel-Based Nonparametric ROC Regression Modelling
}
\description{
The npROCRegression package allows the user to apply in practice the nonparametric induced and direct ROC regression approaches presented in Rodriguez-Alvarez et al. (2011a) and Rodriguez-Alvarez et al. (2011b, 2016) respectively.
}
\details{
\tabular{ll}{
Package: \tab npROCRegression\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-16\cr
License: \tab GPL \cr
}
}
\author{
Maria Xose Rodriguez - Alvarez and Javier Roca-Pardinas

Maintainer: Maria Xose Rodriguez - Alvarez <mxrodriguez@uvigo.es>
}
\references{
Rodriguez-Alvarez, M.X., Roca-Pardinas, J. and Cadarso-Suarez, C. (2011a). ROC curve and covariates: extending induced methodology to the non-parametric framework. Statistics and Computing, 21(4), 483--499.

Rodriguez- Alvarez, M.X., Roca-Pardinas, J. and Cadarso-Suarez, C. (2011b). A new flexible direct ROC regression model - Application to the detection of cardiovascular risk factors by anthropometric measures. Computational Statistics and Data Analysis, 55(12), 3257--3270.

Rodriguez- Alvarez, M.X., Roca-Pardinas, J. and Cadarso-Suarez, C. (2016). Boostrap-based procedures for inference in nonparametric ROC regression analysis. Technical report.
}
\keyword{package}
%\seealso{}
%\examples{}
